requirements.txt
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
cbor==1.0.0
cbor2==5.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
chest==0.2.3
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cwl-eval==1.0.12
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
exceptiongroup==1.2.1
executing==2.0.1
fastparquet==2024.2.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
HeapDict==1.0.1
huggingface-hub==0.23.0
idna==3.7
ijson==3.2.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
inscriptis==2.5.0
ipykernel==6.29.4
ipython==8.18.1
ir_datasets==0.5.7
ir_measures==0.3.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.2.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matchpy==0.5.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multiset==2.1.1
nest-asyncio==1.6.0
networkx==3.2.1
nptyping==1.4.4
numba==0.59.1
numpy==1.26.4
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pandas-stubs==2.2.2.240514
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.18.0
pyjnius==1.6.1
pyparsing==3.1.2
pyterrier-t5 @ git+https://github.com/terrierteam/pyterrier_t5.git@63756ebc2968ab03f46a61f0b391e27873226d75
python-dateutil==2.9.0.post0
python-terrier==0.10.1
pytrec_eval==0.5
pytrec_eval-terrier==0.5.6
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ranx==0.3.19
regex==2024.5.15
repro-eval==0.4.0
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tabulate==0.9.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.40.2
trec-car-tools==2.6
types-pytz==2024.1.0.20240417
typing_extensions==4.11.0
typish==1.9.3
tzdata==2024.1
unlzw3==0.2.2
urllib3==2.2.1
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
wget==3.2
wrapt==1.16.0
zipp==3.18.2
zlib-state==0.1.6
